ecc-openlxp-xds/requirements.txt
bleach~=6.0.0

boto3~=1.40.23

coverage>=5.5,<6.0

django_elasticsearch_dsl_drf>=0.22.5,<0.23.0

django-admin-interface>=0.28.3

django-colorfield>=0.11.0

django-cors-headers>=3.7.0,<3.8.0

django-csp>=3.7,<3.8

django-elasticsearch-dsl>=7.2.2,<9.0

django-model-utils>=4.1.1,<4.2.0

django-mysql>=4.4.0,<5.0.0

django-notifications-hq>=1.8.3, <2.0

Django>=4.2.21,<5.0

djangorestframework>=3.15.2,<3.16.0

elasticsearch>=7.17.0,<7.18.0

flake8>=3.8.4,<3.9.0

gunicorn>=23.0.0, <23.1.0

isort==5.11.5

martor>=1.5.8,<1.6.0

mysql-connector-python>=9.1.0, <9.2.0

openlxp-P1-notification>=0.0.18, <2.0.0

p1-auth>=0.1.6,<1.0.0

PyJWT>=2.0.0

pylint-django==2.5.3

python-slugify>=8.0.1

setuptools==80.9.0

text-unidecode>=1.3
